fix(autostart): make quadlet ownership unforgeable by a compose label
I found that a compose file's own labels: are rendered into the same
[Section] as the trusted Label=podup.project= stamp, in the identical
Label=key=value shape. A service under app-extra declaring
labels: {podup.project: app} produces a unit whose first
Label=podup.project= line reads app (forged) and whose second reads
app-extra (the real stamp). unit_owner took the first match, so
`uninstall -p app` deleted app-extra's unit too, reopening the
cross-project-deletion bug through a different door than the
filename-prefix one this branch already closed.

I kept the Label=podup.project= line as-is (Podman still needs it for
container/secret scoping) and instead had every unit builder
(container/build/network/volume) emit a `# podup-owner: <project>`
comment as the literal first line of the generated unit. systemd
ignores # comments, and compose labels can only ever render as
Label=key=value entries, never as a comment line, so this marker
can't be forged the same way. unit_owner and installed_units now read
ownership from that marker instead of the label; a unit with no
marker is treated the same as before (unproven, skipped, warned about,
never deleted).

I split quadlet.rs's test module into a sibling internal/autostart/
quadlet/tests.rs to stay under the file-length limit, and added a test
that reproduces the forged-label attack directly (confirmed it fails
against the prior label-based check and passes now).
